fpspreadsheet: Add unit tests for some more page layout properties (Fit width/height to pages, start page number) (completing the first 1000 test cases...)

+++ components/fpspreadsheet/fpstypes.pas
@@ -688,7 +688,7 @@ type
     HeaderMargin: Double;
     FooterMargin: Double;
     StartPageNumber: Integer;
-    ScalingFactor: Double;  // in percent
+    ScalingFactor: Integer;  // in percent
     FitWidthToPages: Integer;
     FitHeightToPages: Integer;
     Copies: Integer;
